requirements.txt
pyyaml
psycopg2-binary
sqlalchemy
requests
pandas